requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.3
diff-match-patch==20241021
django-cors-headers==4.7.0
django-import-export==4.3.7
django-ninja==1.4.1
django-ninja-extra==0.30.0
django-ninja-jwt==5.3.7
django-unfold==0.57.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
injector==0.22.0
numpy==2.0.0
oauth2client==4.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
